Backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.2
numpy<1.24,>=1.22
tensorflow==2.12.0
scikit-learn>=1.2.0
joblib==1.2.0
Pillow==9.5.0
python-multipart==0.0.6
h5py>=3.7.0